OH putain frerot j'ai réussi sample décoder je veux mourir de debugg faut regarder avec tristan poour isfull et merde yes quoi j'ai eu SamplesDecoder et il reste un peu de next message mais ca devrait le faire

+++ src/ch/epfl/javions/demodulation/PowerWindow.java
@@ -13,7 +13,7 @@ public final class PowerWindow {
     private int windowSize;
     private int position;
     private PowerComputer computer;
-    private int sizeB;
+    private long sizeB;
     private int[] batchpowerOne;
     private int[] batchpowerTwo;
 
@@ -97,6 +97,7 @@ public final class PowerWindow {
         if(position%Math.scalb(1d, 16) == 0){
             batchpowerOne = batchpowerTwo;
         }
+
     }
 
     /**
